Added Logger,Comments and Response Entity.

+++ src/main/java/com/cg/vegetable/service/VegetableServiceImpl.java
@@ -1,29 +1,43 @@
 package com.cg.vegetable.service;
 
 import java.util.List;
+import java.util.Optional;
+
+import org.apache.logging.log4j.LogManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.cg.vegetable.module.Vegetable;
-import com.cg.vegetable.repository.ICartRepository;
-import com.cg.vegetable.repository.ICustomerRepository;
 import com.cg.vegetable.repository.IVegetableRepository;
 
 @Service
 public class VegetableServiceImpl implements IVegetableService {
+
+	/**
+	 * Logger
+	 */
+
+	org.apache.logging.log4j.Logger logger = LogManager.getLogger(VegetableServiceImpl.class);
+
+	/**
+	 * AutoWiring the repository class to call down the service
+	 */
 	@Autowired
 	IVegetableRepository vegrepo;
 
-	@Autowired
-	ICartRepository cartrepo;
-
-	@Autowired
-	ICustomerRepository custrepo;
-
+	/**
+	 * This below function is used to create a new vegetable and redirects to the
+	 * vegetable repository
+	 */
 	@Override
 	public Vegetable save(Vegetable dto) {
+		logger.info("Successfully Added" + dto);
 		return vegrepo.save(dto);
 	}
+
+	/**
+	 * This below function is used to update a specific vegetable based on the give
+	 * vegId and redirects to the vegetable repository
+	 */
 
 	@Override
 	public Vegetable update(int vegId, Vegetable dto) {
@@ -34,37 +48,68 @@ public class VegetableServiceImpl implements IVegetableService {
 		v.setPrice(dto.getPrice());
 		v.setType(dto.getType());
 		v.setQuantity(dto.getQuantity());
-
+		logger.info("Update Successfully" + v);
 		return vegrepo.save(v);
 	}
 
+	/**
+	 * This below function is used to delete a specific vegetable based on the given
+	 * vegId and redirects to the vegetable repository
+	 */
 	@Override
 	public Vegetable deleteByvegId(int vegId) {
-		Vegetable veg = vegrepo.findById(vegId).get();
+		Optional<Vegetable> opt = vegrepo.findById(vegId);
+		if (!opt.isPresent()) {
+			return null;
+		}
 		vegrepo.deleteById(vegId);
-		return veg;
+		return opt.get();
 	}
+
+	/**
+	 * This below function is used to get a specific vegetable and vegId as
+	 * parameter and redirects to the vegetable repository
+	 */
 
 	@Override
 	public Vegetable viewVegetableById(int vegId) {
-		Vegetable vgd = vegrepo.findById(vegId).get();
-		return vgd;
+		Optional<Vegetable> opt = vegrepo.findById(vegId);
+		if (!opt.isPresent()) {
+			return null;
+		}
+		logger.info(" Viewed By Id Successfully " + opt);
+		return opt.get();
 
 	}
 
+	/**
+	 * This below function is used to get all the vegetables and redirects to the
+	 * vegetable repository
+	 */
 	@Override
 	public List<Vegetable> retrive() {
+		logger.info("Viewed Successfully");
 		return vegrepo.findAll();
 	}
 
+	/**
+	 * This below function is used to get a specific vegetable and type as parameter
+	 * and redirects to the vegetable repository
+	 */
 	@Override
 	public List<Vegetable> viewVegetableList(String type) {
-
+		logger.info("Viewed Successfully" + type);
 		return vegrepo.findVegetableList(type);
 	}
 
+	/**
+	 * This below function is used to get a specific vegetable and name as parameter
+	 * and redirects to the vegetable repository
+	 */
+
 	@Override
 	public List<Vegetable> viewVegetableByName(String name) {
+		logger.info("Viewed Vegetable By Name Successfully" + name);
 		return vegrepo.findVegetableByName(name);
 
 	}
